Add the address field for warp.
#1385
In some cases, the outbound configuration for wireguard requires the addition of the 'address' field in order to establish a proper connection. 
It should be noted that the absence of the 'address' field configuration results in the inability to establish connections via domain names, only IP addresses can be used for connections.